move openai functions to modular folder
